Fix encoding issue
Fixes #7 using a Blob rather than btoa (as btoa() creates a Base64-encoded ASCII string).

Side modifiation, the {{VALUES}}\n substitution does not works on my environnement as the substituion is searching for a litteral \n in the template. Feel free to remove this modification if it's not relevant and I juste missed something along the way :-)