Decrypt the secret when the user hits the Enter key
Thanks for this great tool!

I found that my natural inclination was to type my password and then hit the Enter key, and I kept expecting that to trigger decryption. It looks like there's no handling for the Enter key, so I thought I'd add it as a feature to see if you find it useful.